src/requirements.txt
llama_cpp_python==0.3.20
llama-index-core==0.14.20
llama-index-llms-llama-cpp==0.6.0
llama-index-embeddings-huggingface==0.7.0
llama-index-graph-stores-neo4j==0.7.0

aiohttp==3.9.0
fastapi==0.110.0
python-dotenv==1.0.1
uvicorn==0.27.1